refactor(gui): extract shared overlay library
