+++ crates/openlogi-gui/src/action_icons.rs
@@ -2,7 +2,7 @@
 
 use openlogi_core::binding::{Action, ActionRingIcon};
 
-use crate::action_ring_icons::ring_icon_path;
+use openlogi_gui::action_ring_icons::ring_icon_path;
 
 /// Embedded Lucide asset for an action.
 pub(crate) fn action_icon_path(action: &Action) -> &'static str {
