+++ crates/openlogi-gui/src/components/action_ring_panel.rs
@@ -15,10 +15,10 @@ use openlogi_core::binding::{ActionRingEntry, ActionRingLayout, ActionRingSlot};
 
 use self::editor::action_library;
 use crate::action_icons::action_icon_path;
-use crate::action_ring_geometry::slot_offset;
-use crate::action_ring_icons::ring_icon_path;
 use crate::state::AppState;
 use crate::theme::{self, Palette, Typography as _};
+use openlogi_gui::action_ring_geometry::slot_offset;
+use openlogi_gui::action_ring_icons::ring_icon_path;
 
 /// Stateful Actions Ring editor. Ring configuration itself lives in
 /// [`AppState`]; this entity owns selection and editor input state.
