+++ crates/openlogi-gui/src/locale.rs
@@ -1,8 +1,25 @@
-//! Pure locale negotiation shared by the settings app and overlay helper.
+//! Locale negotiation: which shipped catalog a BCP-47 code resolves to.
+//!
+//! Shared by both binaries — the settings app picks the locale from its stored
+//! [`AppSettings`](openlogi_core::config::AppSettings) (see `i18n`), the
+//! overlay helper from the code the agent hands it per invocation. Keeping one
+//! implementation is what stops the two processes from disagreeing about, say,
+//! whether `zh-Hant-HK` is Hong Kong or Taiwan.
+//!
+//! Setting the locale is global: `rust_i18n` stores it in a process-wide atomic
+//! that gpui-component reads too, so [`crate::locale::activate`] re-localizes
+//! the framework's own widget strings alongside ours.
 
 use fluent_langneg::{LanguageIdentifier, NegotiationStrategy, negotiate_languages};
 
-/// Locales the GUI ships, as `(code, native name)` in picker order.
+/// Locales the GUI ships, as `(code, native name)`. The codes match the
+/// `locales/*.yml` filenames; a subset (`en`, `zh-CN`, `zh-HK`, `it`) also
+/// matches gpui-component's bundled `ui.yml`, so choosing one localizes the
+/// framework's own widgets too. Under a locale the framework doesn't bundle, our
+/// app strings localize but gpui-component's built-in widget strings fall back
+/// to English.
+/// Order here is the order shown in the Settings picker (after "Follow system"):
+/// native-name alphabetical within each script.
 pub const SUPPORTED: &[(&str, &str)] = &[
     ("da", "Dansk"),
     ("de", "Deutsch"),
@@ -26,7 +43,9 @@ pub const SUPPORTED: &[(&str, &str)] = &[
     ("ko", "한국어"),
 ];
 
-/// Resolve an explicit setting or system locale to a shipped locale.
+/// Resolve the locale to apply, preferring an explicit stored `setting`, then
+/// the system locale, and finally `"en"`. An unrecognized stored code is
+/// treated as "follow system" rather than failing.
 #[must_use]
 pub fn resolve(setting: Option<&str>) -> &'static str {
     setting
@@ -39,6 +58,17 @@ pub fn resolve(setting: Option<&str>) -> &'static str {
         .unwrap_or("en")
 }
 
+/// Collapse an arbitrary BCP-47 locale onto one of [`SUPPORTED`], or `None`,
+/// by matching its primary subtag. Three families need more than a primary-tag
+/// match:
+/// - `zh` is decided by examining all subtags for script and region: explicit
+///   `Hans` → `zh-CN` (always wins); `hk` / `mo` region → `zh-HK`; `tw` region
+///   or bare `Hant` script → `zh-TW`; no recognized indicator → `zh-CN`. So
+///   `zh-Hans-HK` stays Simplified (script wins), `zh-Hant-HK` resolves to Hong
+///   Kong (region wins over generic script), and bare `zh-Hant` → Taiwan.
+/// - `pt` splits on region: a `br` subtag → `pt-BR`, otherwise `pt-PT`.
+/// - Norwegian's `nb` / `nn` / the macrolanguage `no` all fold onto `nb`
+///   (the catalog ships Bokmål, shown as "Norsk").
 fn match_supported(code: &str) -> Option<&'static str> {
     let requested = code.replace('_', "-").parse::<LanguageIdentifier>().ok()?;
     special_locale(&requested).or_else(|| lookup_supported(&requested))
@@ -81,8 +111,8 @@ fn lookup_supported(requested: &LanguageIdentifier) -> Option<&'static str> {
         NegotiationStrategy::Lookup,
     )
     .into_iter()
-    .next()?
-    .to_string();
+    .next()?;
+    let matched = matched.to_string();
     SUPPORTED
         .iter()
         .find_map(|(code, _)| (*code == matched).then_some(*code))
@@ -95,6 +125,17 @@ fn supported_langids() -> Vec<LanguageIdentifier> {
         .collect()
 }
 
+/// Switch the process-global locale to the resolution of `language`
+/// (`None` = follow system). The single resolve→`set_locale` surface for every
+/// caller — app startup, the live Settings switch, and each overlay
+/// invocation — so the resolution policy can't drift between them.
+///
+/// The caller is responsible for refreshing any open window afterwards; views
+/// already rendered do not re-read the locale on their own.
+pub fn activate(language: Option<&str>) {
+    rust_i18n::set_locale(resolve(language));
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -102,22 +143,40 @@ mod tests {
     #[test]
     fn maps_locale_variants() {
         assert_eq!(match_supported("zh-Hans-CN"), Some("zh-CN"));
+        assert_eq!(match_supported("zh-CN"), Some("zh-CN"));
         assert_eq!(match_supported("zh-Hans-HK"), Some("zh-CN"));
         assert_eq!(match_supported("zh-Hant-TW"), Some("zh-TW"));
+        assert_eq!(match_supported("zh-TW"), Some("zh-TW"));
         assert_eq!(match_supported("zh-Hant"), Some("zh-TW"));
         assert_eq!(match_supported("zh-HK"), Some("zh-HK"));
         assert_eq!(match_supported("zh-Hant-HK"), Some("zh-HK"));
+        assert_eq!(match_supported("ja"), Some("ja"));
         assert_eq!(match_supported("ja-JP"), Some("ja"));
+        assert_eq!(match_supported("ru"), Some("ru"));
         assert_eq!(match_supported("ru-RU"), Some("ru"));
         assert_eq!(match_supported("en-US"), Some("en"));
+        assert_eq!(match_supported("it"), Some("it"));
         assert_eq!(match_supported("it-IT"), Some("it"));
         assert_eq!(match_supported("fr-FR"), Some("fr"));
+        assert_eq!(match_supported("de"), Some("de"));
         assert_eq!(match_supported("ko-KR"), Some("ko"));
         assert_eq!(match_supported("pt"), Some("pt-PT"));
+        assert_eq!(match_supported("pt-PT"), Some("pt-PT"));
         assert_eq!(match_supported("pt-BR"), Some("pt-BR"));
         assert_eq!(match_supported("nb-NO"), Some("nb"));
         assert_eq!(match_supported("no"), Some("nb"));
         assert_eq!(match_supported("nn"), Some("nb"));
         assert_eq!(match_supported("klingon"), None);
+    }
+
+    #[test]
+    fn explicit_setting_wins_over_system() {
+        assert_eq!(resolve(Some("zh-CN")), "zh-CN");
+        // An unknown stored code falls through to system/`en`, never panics.
+        assert!(
+            SUPPORTED
+                .iter()
+                .any(|(c, _)| *c == resolve(Some("klingon")))
+        );
     }
 }
