+++ crates/openlogi-gui/src/main.rs
@@ -30,10 +30,7 @@ macro_rules! tr {
 }
 
 mod action_icons;
-mod action_ring_geometry;
-mod action_ring_icons;
 mod app;
-mod app_assets;
 mod app_menu;
 mod asset;
 mod components;
@@ -177,10 +174,10 @@ fn main() -> Result<()> {
     let (asset_ctrl_tx, mut asset_ctrl_rx) = tokio::sync::mpsc::unbounded_channel::<AssetCommand>();
     let asset_ctrl_self_tx = asset_ctrl_tx.clone();
 
-    // `with_assets` registers the embedded app logo ([`app_assets`]) plus the
-    // lucide SVGs that back `gpui_component::IconName`; without it `img()` /
-    // `Icon` would fail to load.
-    let app = gpui_platform::application().with_assets(app_assets::AppAssets);
+    // `with_assets` registers the embedded app logo
+    // ([`openlogi_gui::app_assets`]) plus the lucide SVGs that back
+    // `gpui_component::IconName`; without it `img()` / `Icon` would fail to load.
+    let app = gpui_platform::application().with_assets(openlogi_gui::app_assets::AppAssets);
 
     // URL scheme: `open openlogi://open-settings` from the agent's tray or
     // external apps. Works for both cold start (macOS launches the app then
