+++ crates/openlogi-gui/src/windows/settings/language.rs
@@ -35,7 +35,7 @@ pub(super) fn language_options() -> Vec<LanguageOption> {
         localize_label: true,
     }];
     options.extend(
-        crate::i18n::SUPPORTED
+        openlogi_gui::locale::SUPPORTED
             .iter()
             .map(|(code, name)| LanguageOption {
                 label: name,
@@ -59,7 +59,7 @@ pub(super) fn selected_language_index(
 }
 
 /// The language picker field. "Follow system" clears the stored preference
-/// (`None`); explicit locale entries come from [`crate::i18n::SUPPORTED`].
+/// (`None`); explicit locale entries come from [`openlogi_gui::locale::SUPPORTED`].
 #[allow(
     clippy::needless_pass_by_value,
     reason = "built inside an `Fn` render closure, so a `&Entity` parameter would make \
