+++ crates/openlogi-gui/src/windows/settings/updates.rs
@@ -146,7 +146,7 @@ fn update_hero(updater: &Entity<Updater>, pal: Palette, cx: &mut App) -> AnyElem
                 .gap_3()
                 .flex_1()
                 .min_w_0()
-                .child(img(crate::app_assets::LOGO).w(px(52.)).h(px(52.)))
+                .child(img(openlogi_gui::app_assets::LOGO).w(px(52.)).h(px(52.)))
                 .child(
                     v_flex()
                         .gap_1()
